Refactor: check dto size length (#219)
