+++ src/main/java/com/gaduationproject/cre8/app/community/dto/response/CommunityBoardResponseDto.java
@@ -1,4 +1,4 @@
-package com.gaduationproject.cre8.adminapp.communityBoard.dto;
+package com.gaduationproject.cre8.app.community.dto.response;
 
 import com.gaduationproject.cre8.domain.community.entity.CommunityBoard;
 import lombok.AccessLevel;
